cap_model/requirements.txt

numpy>=1.24.3,<2.0.0
pandas>=2.0.3,<3.0.0
scipy>=1.11.0,<2.0.0

# Machine Learning - Scikit-learn
scikit-learn>=1.3.0,<2.0.0

# Gradient Boosting Libraries
xgboost>=1.7.6,<3.0.0
lightgbm>=4.0.0,<5.0.0
catboost>=1.2.0,<2.0.0

# Data Processing & Imbalance Handling
imbalanced-learn>=0.11.0,<1.0.0

# Visualization
matplotlib>=3.7.2,<4.0.0
seaborn>=0.12.2,<1.0.0

# MLflow Integration - Experiment Tracking
mlflow==2.9.2

# Model Persistence
joblib>=1.3.1,<2.0.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Data Source - Kaggle (for downloading datasets)
kaggle>=1.5.0
hatchling>=1.15.1

# Logging
loguru>=0.7.0
